+++ Packages/idv.jlchntoz.vvmw/Runtime/VVMW/FrontendHandler_PlayList.cs
@@ -147,6 +147,7 @@ namespace JLChnToZ.VRC.VVMW {
             }
             core.PlayUrl(playListUrls[localPlayingIndex], playListUrlsQuest[localPlayingIndex], playListPlayerIndex[localPlayingIndex]);
             core.SetTitle(playListEntryTitles[localPlayingIndex], playListTitles[localPlayListIndex - 1]);
+            localCurrentTitle = "";
             RequestSync();
             UpdateState();
         }